+++ src/ui/main_deck.rs
@@ -12,7 +12,7 @@ pub fn MainDeck() -> Element {
         div {
             class: "relative bg-slate-50/2 size-[9vw] aspect-square flex items-center justify-center border-0.5",
             CardStack {
-                length: (state.main_deck_length)(),
+                length: (state.main_deck_length)() as usize,
                 image_url: CARD_BACK,
             }
         }
